fixing docs and Trigger name
